LG-10301: Move and rename AddressSearch component. (#9100)
* Move file, re-export

* changelog: Internal, In-Person Proofing, Move address search input component into its own file

* Rename type; remove leftover exports

* Fix naming in spec
